amy help!
Amy, please help me with the ghost padding we were discussing in class.
It looks like it's an issue on the left side of the containers.  Also,
could you please look into the header issue where it cuts off the logo?
Thank you so much in advance for your help!